added git reset
